{{cookiecutter.project_slug}}/requirements/base.txt
flask==1.0.2
blinker==1.4
environs==4.1.0
simplejson==3.16.0

psycopg2-binary==2.7.7

flask-talisman==0.6.0

flask-sqlalchemy==2.3.2
flask-migrate==2.4.0
flask-caching==1.7.0
flask-wtf==0.14.2

flask-security==3.0.0

flask-babel==0.12.2

whitenoise==4.1.2

flask-marshmallow==0.10.0
marshmallow-sqlalchemy==0.16.1

flask-apispec==0.8.0
